chore: refactor to operation and state folders
