+++ f2soperator/operation/dispatcher/eventhandler.go
@@ -1,0 +1,11 @@
+package dispatcher
+
+import (
+	"butschi84/f2s/state/eventmanager"
+	"fmt"
+)
+
+func handleEvents(event eventmanager.Event) {
+	logging.Info("processing event", fmt.Sprintf("'%s'", string(event.Type)))
+
+}
