+++ f2soperator/operation/dispatcher/types.go
@@ -1,8 +1,8 @@
 package dispatcher
 
 import (
-	typesV1alpha1 "butschi84/f2s/configuration/api/types/v1alpha1"
 	"butschi84/f2s/hub"
+	typesV1alpha1 "butschi84/f2s/state/configuration/api/types/v1alpha1"
 	"sync"
 )
