+++ f2soperator/operation/metrics/eventhandler.go
@@ -1,8 +1,8 @@
 package metrics
 
 import (
-	"butschi84/f2s/eventmanager"
 	"butschi84/f2s/services/prometheus"
+	"butschi84/f2s/state/eventmanager"
 	"fmt"
 	"time"
 )
