+++ f2soperator/operation/metrics/main.go
