+++ f2soperator/operation/routes/configuration.go
