+++ f2soperator/operation/routes/endpoints.go
