+++ f2soperator/state/configuration/api/clientset/v1alpha1/functions.go
@@ -1,7 +1,7 @@
 package v1alpha1
 
 import (
-	"butschi84/f2s/configuration/api/types/v1alpha1"
+	"butschi84/f2s/state/configuration/api/types/v1alpha1"
 	"context"
 	"fmt"
