+++ f2soperator/state/configuration/api/types/v1alpha1/register.go
